Add styled-components, make some ugly buttons
