+++ package.json
@@ -47,17 +47,19 @@
   },
   "license": "ISC",
   "dependencies": {
-    "electron-store": "^6.0.1",
     "cross-env": "^7.0.3",
     "electron-is-dev": "^2.0.0",
     "electron-root-path": "^1.0.16",
+    "electron-store": "^6.0.1",
     "electron-updater": "^4.3.5",
     "execa": "^5.1.1",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
+    "react-router-dom": "^5.3.0",
     "react-scripts": "^4.0.3",
     "sass": "^1.40.1",
     "shell-path": "2.1.0",
+    "styled-components": "^5.3.1",
     "wait-on": "^6.0.0",
     "web-vitals": "^2.1.0"
   },
